ditto

+++ src/ATen/native/xpu/sycl/UpSampleNearest3dKernels.cpp
@@ -352,7 +352,7 @@ void upsample_nearest3d_backward_template(
           width_scale,
           index_bw_op);
   auto work_group_size = syclMaxWorkGroupSize(kfn);
-  int64_t work_group_num = at::ceil_div(n, (int64_t)work_group_size);
+  int64_t work_group_num = at::ceil_div(n, work_group_size);
   sycl_kernel_submit(
       work_group_num * work_group_size, work_group_size, queue, kfn);
 }
